# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gig Trail Tour Calculator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Profile {
  id: number;
  name: string;
  actType: string;
  /** @nullable */
  homeBase: string | null;
  peopleCount: number;
  /** @nullable */
  defaultVehicleId: number | null;
  avgAccomPerNight: number;
  avgFoodPerDay: number;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CreateProfileBody {
  name: string;
  actType: string;
  /** @nullable */
  homeBase?: string | null;
  peopleCount: number;
  /** @nullable */
  defaultVehicleId?: number | null;
  avgAccomPerNight: number;
  avgFoodPerDay: number;
  /** @nullable */
  notes?: string | null;
}

export interface Vehicle {
  id: number;
  name: string;
  fuelType: string;
  avgConsumption: number;
  /** @nullable */
  maxPassengers: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CreateVehicleBody {
  name: string;
  fuelType: string;
  avgConsumption: number;
  /** @nullable */
  maxPassengers?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface Run {
  id: number;
  /** @nullable */
  profileId: number | null;
  /** @nullable */
  vehicleId: number | null;
  /** @nullable */
  origin: string | null;
  /** @nullable */
  destination: string | null;
  distanceKm: number;
  returnTrip: boolean;
  fuelPrice: number;
  showType: string;
  /** @nullable */
  fee: number | null;
  /** @nullable */
  capacity: number | null;
  /** @nullable */
  ticketPrice: number | null;
  /** @nullable */
  expectedAttendancePct: number | null;
  /** @nullable */
  dealType: string | null;
  /** @nullable */
  splitPct: number | null;
  /** @nullable */
  guarantee: number | null;
  /** @nullable */
  merchEstimate: number | null;
  /** @nullable */
  marketingCost: number | null;
  /** @nullable */
  accommodationCost: number | null;
  /** @nullable */
  foodCost: number | null;
  /** @nullable */
  extraCosts: number | null;
  /** @nullable */
  totalCost: number | null;
  /** @nullable */
  totalIncome: number | null;
  /** @nullable */
  totalProfit: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CreateRunBody {
  /** @nullable */
  profileId?: number | null;
  /** @nullable */
  vehicleId?: number | null;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  destination?: string | null;
  distanceKm: number;
  returnTrip: boolean;
  fuelPrice: number;
  showType: string;
  /** @nullable */
  fee?: number | null;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  ticketPrice?: number | null;
  /** @nullable */
  expectedAttendancePct?: number | null;
  /** @nullable */
  dealType?: string | null;
  /** @nullable */
  splitPct?: number | null;
  /** @nullable */
  guarantee?: number | null;
  /** @nullable */
  merchEstimate?: number | null;
  /** @nullable */
  marketingCost?: number | null;
  /** @nullable */
  accommodationCost?: number | null;
  /** @nullable */
  foodCost?: number | null;
  /** @nullable */
  extraCosts?: number | null;
  /** @nullable */
  totalCost?: number | null;
  /** @nullable */
  totalIncome?: number | null;
  /** @nullable */
  totalProfit?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface Tour {
  id: number;
  name: string;
  /** @nullable */
  profileId: number | null;
  /** @nullable */
  vehicleId: number | null;
  /** @nullable */
  startLocation: string | null;
  /** @nullable */
  endLocation: string | null;
  returnHome: boolean;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  endDate: string | null;
  /** @nullable */
  defaultFoodCost: number | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  totalDistance: number | null;
  /** @nullable */
  totalCost: number | null;
  /** @nullable */
  totalIncome: number | null;
  /** @nullable */
  totalProfit: number | null;
  createdAt: string;
}

export interface CreateTourBody {
  name: string;
  /** @nullable */
  profileId?: number | null;
  /** @nullable */
  vehicleId?: number | null;
  /** @nullable */
  startLocation?: string | null;
  /** @nullable */
  endLocation?: string | null;
  returnHome?: boolean;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  defaultFoodCost?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface TourStop {
  id: number;
  tourId: number;
  stopOrder: number;
  /** @nullable */
  date: string | null;
  city: string;
  /** @nullable */
  venueName: string | null;
  showType: string;
  /** @nullable */
  fee: number | null;
  /** @nullable */
  capacity: number | null;
  /** @nullable */
  ticketPrice: number | null;
  /** @nullable */
  expectedAttendancePct: number | null;
  /** @nullable */
  dealType: string | null;
  /** @nullable */
  splitPct: number | null;
  /** @nullable */
  guarantee: number | null;
  /** @nullable */
  merchEstimate: number | null;
  /** @nullable */
  marketingCost: number | null;
  /** @nullable */
  accommodationCost: number | null;
  /** @nullable */
  extraCosts: number | null;
  /** @nullable */
  distanceOverride: number | null;
  /** @nullable */
  fuelPriceOverride: number | null;
  /** @nullable */
  notes: string | null;
}

export interface CreateTourStopBody {
  stopOrder?: number;
  /** @nullable */
  date?: string | null;
  city: string;
  /** @nullable */
  venueName?: string | null;
  showType: string;
  /** @nullable */
  fee?: number | null;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  ticketPrice?: number | null;
  /** @nullable */
  expectedAttendancePct?: number | null;
  /** @nullable */
  dealType?: string | null;
  /** @nullable */
  splitPct?: number | null;
  /** @nullable */
  guarantee?: number | null;
  /** @nullable */
  merchEstimate?: number | null;
  /** @nullable */
  marketingCost?: number | null;
  /** @nullable */
  accommodationCost?: number | null;
  /** @nullable */
  extraCosts?: number | null;
  /** @nullable */
  distanceOverride?: number | null;
  /** @nullable */
  fuelPriceOverride?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface TourWithStops {
  id: number;
  name: string;
  /** @nullable */
  profileId: number | null;
  /** @nullable */
  vehicleId: number | null;
  /** @nullable */
  startLocation: string | null;
  /** @nullable */
  endLocation: string | null;
  returnHome: boolean;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  endDate: string | null;
  /** @nullable */
  defaultFoodCost: number | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  totalDistance: number | null;
  /** @nullable */
  totalCost: number | null;
  /** @nullable */
  totalIncome: number | null;
  /** @nullable */
  totalProfit: number | null;
  createdAt: string;
  stops: TourStop[];
}

export interface DashboardSummary {
  totalRuns: number;
  totalTours: number;
  totalProfiles: number;
  totalVehicles: number;
  totalKmDriven: number;
  totalIncome: number;
  totalProfit: number;
  worthTheDrive: number;
  tightMargins: number;
  notWorthIt: number;
}

export interface DashboardRecent {
  recentRuns: Run[];
  recentTours: Tour[];
}