# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gig Trail Tour Calculator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: getProfiles
      tags: [profiles]
      summary: List all profiles
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /runs:
    get:
      operationId: getRuns
      tags: [runs]
      summary: List all single show calculations
      responses:
        "200":
          description: List of runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
    post:
      operationId: createRun
      tags: [runs]
      summary: Create a single show calculation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunBody"
      responses:
        "201":
          description: Created run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
  /runs/{id}:
    get:
      operationId: getRun
      tags: [runs]
      summary: Get a run by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRun
      tags: [runs]
      summary: Update a run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunBody"
      responses:
        "200":
          description: Updated run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
    delete:
      operationId: deleteRun
      tags: [runs]
      summary: Delete a run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Profile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        actType:
          type: string
        homeBase:
          type: ["string", "null"]
        peopleCount:
          type: integer
        defaultVehicleId:
          type: ["integer", "null"]
        avgAccomPerNight:
          type: number
        avgFoodPerDay:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - actType
        - homeBase
        - peopleCount
        - defaultVehicleId
        - avgAccomPerNight
        - avgFoodPerDay
        - notes
        - createdAt
    CreateProfileBody:
      type: object
      properties:
        name:
          type: string
        actType:
          type: string
        homeBase:
          type: ["string", "null"]
        peopleCount:
          type: integer
        defaultVehicleId:
          type: ["integer", "null"]
        avgAccomPerNight:
          type: number
        avgFoodPerDay:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - name
        - actType
        - peopleCount
        - avgAccomPerNight
        - avgFoodPerDay
...
    Run:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        origin:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        distanceKm:
          type: number
        returnTrip:
          type: boolean
        fuelPrice:
          type: number
        showType:
          type: string
        fee:
          type: ["number", "null"]
        capacity:
          type: ["integer", "null"]
        ticketPrice:
          type: ["number", "null"]
        expectedAttendancePct:
          type: ["number", "null"]
        dealType:
          type: ["string", "null"]
        splitPct:
          type: ["number", "null"]
        guarantee:
          type: ["number", "null"]
        merchEstimate:
          type: ["number", "null"]
        marketingCost:
          type: ["number", "null"]
        accommodationCost:
          type: ["number", "null"]
        foodCost:
          type: ["number", "null"]
        extraCosts:
          type: ["number", "null"]
        totalCost:
          type: ["number", "null"]
        totalIncome:
          type: ["number", "null"]
        totalProfit:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - profileId
        - vehicleId
        - origin
        - destination
        - distanceKm
        - returnTrip
        - fuelPrice
        - showType
        - fee
        - capacity
        - ticketPrice
        - expectedAttendancePct
        - dealType
        - splitPct
        - guarantee
        - merchEstimate
        - marketingCost
        - accommodationCost
        - foodCost
        - extraCosts
        - totalCost
        - totalIncome
        - totalProfit
        - notes
        - createdAt
    CreateRunBody:
      type: object
      properties:
        profileId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        origin:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        distanceKm:
          type: number
        returnTrip:
          type: boolean
        fuelPrice:
          type: number
        showType:
          type: string
        fee:
          type: ["number", "null"]
        capacity:
          type: ["integer", "null"]
        ticketPrice:
          type: ["number", "null"]
        expectedAttendancePct:
          type: ["number", "null"]
        dealType:
          type: ["string", "null"]
        splitPct:
          type: ["number", "null"]
        guarantee:
          type: ["number", "null"]
        merchEstimate:
          type: ["number", "null"]
        marketingCost:
          type: ["number", "null"]
        accommodationCost:
          type: ["number", "null"]
        foodCost:
          type: ["number", "null"]
        extraCosts:
          type: ["number", "null"]
        totalCost:
          type: ["number", "null"]
        totalIncome:
          type: ["number", "null"]
        totalProfit:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - distanceKm
        - returnTrip
        - fuelPrice
        - showType
...