# lib/api-zod/src/generated/types/createRunBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gig Trail Tour Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRunBody {
  /** @nullable */
  profileId?: number | null;
  /** @nullable */
  vehicleId?: number | null;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  destination?: string | null;
  distanceKm: number;
  returnTrip: boolean;
  fuelPrice: number;
  showType: string;
  /** @nullable */
  fee?: number | null;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  ticketPrice?: number | null;
  /** @nullable */
  expectedAttendancePct?: number | null;
  /** @nullable */
  dealType?: string | null;
  /** @nullable */
  splitPct?: number | null;
  /** @nullable */
  guarantee?: number | null;
  /** @nullable */
  merchEstimate?: number | null;
  /** @nullable */
  marketingCost?: number | null;
  /** @nullable */
  accommodationCost?: number | null;
  /** @nullable */
  foodCost?: number | null;
  /** @nullable */
  extraCosts?: number | null;
  /** @nullable */
  totalCost?: number | null;
  /** @nullable */
  totalIncome?: number | null;
  /** @nullable */
  totalProfit?: number | null;
  /** @nullable */
  notes?: string | null;
}