# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gig Trail Tour Calculator API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  actType: string;
  /** @nullable */
  homeBase: string | null;
  peopleCount: number;
  /** @nullable */
  defaultVehicleId: number | null;
  avgAccomPerNight: number;
  avgFoodPerDay: number;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}